Comprehensive Process Documentation
Raw Data to Insights:
This commit includes the comprehensive process documentation file, named "Processes and Observations". It encompasses all steps undertaken in this project, starting from the initial download of raw data, utilization of VBA scripts for preliminary processing, to the SQL queries and associated observations. This document serves as a detailed record of the project's progression, offering valuable insights into the data analysis process.